requirements.txt
pandas==2.0.2
pybit==5.6.2
pyTelegramBotAPI==4.14.0
PyYAML==6.0
tabulate==0.9.0